finish splitting spawn, move test/utils into a single file
