+++ test/spawn/caught.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should ignore caught errors', t => {
   spawn('catch-no-such-module.js', out => {
