+++ test/spawn/conceal.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should conceal the wrapper', t => {
   // require.main should be main.js not wrap.js!
