+++ test/spawn/exit-code.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should pass through the exit code', t => {
   spawn('exit.js').on('exit', code => {
