+++ test/spawn/expose-gc.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should pass unknown args to node binary', t => {
   spawn('--expose_gc gc.js foo', out => {
