+++ test/spawn/kill-fork.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should kill the forked processes', t => {
   spawn('pid.js', out => {
