+++ test/spawn/node-env.js
@@ -1,6 +1,6 @@
 const tap = require('tap');
 
-const spawn = require('../utils/spawn');
+const { spawn } = require('../utils');
 
 tap.test('should *not* set NODE_ENV', t => {
   spawn('env.js', out => {
